int -> float во всех методах StatsdExporterClientInterface

+++ src/MetricCollector.php
@@ -72,7 +72,7 @@ abstract class MetricCollector implements LoggerAwareInterface
     }
 
 
-    final protected function count(string $key, int $value, array $tags = [])
+    final protected function count(string $key, float $value, array $tags = [])
     {
         $this->collectedMetrics[] = ['count' => [$key, $value, $tags]];
     }
@@ -96,13 +96,13 @@ abstract class MetricCollector implements LoggerAwareInterface
     }
 
 
-    final protected function gauge(string $key, int $value, array $tags = [])
+    final protected function gauge(string $key, float $value, array $tags = [])
     {
         $this->collectedMetrics[] = ['gauge' => [$key, $value, $tags]];
     }
 
 
-    final protected function summary(string $key, int $value, array $tags = [])
+    final protected function summary(string $key, float $value, array $tags = [])
     {
         $this->collectedMetrics[] = ['summary' => [$key, $value, $tags]];
     }
